--- app/tailwind.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-    content: [
-      './app/**/*.{js,jsx,ts,tsx}',
-      './components/**/*.{js,jsx,ts,tsx}',
-      './services/**/*.{js,jsx,ts,tsx}'
-    ],
-    theme: { extend: {} },
-    plugins: [],
-  };
